ChrysisRagDeepThought_final/chrysisRagDeepThought_pre-run/requirements.txt
numpy
pandas
scikit-learn
together
tqdm
openpyxl
faiss-cpu
scipy
sentence-transformers
llama-index-core
matplotlib 
seaborn  